requirements.txt
astroid==3.3.10
black==25.1.0
click==8.2.1
dill==0.4.0
flake8==7.3.0
iniconfig==2.1.0
isort==6.0.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.7
pytest==8.4.1
PyYAML==6.0.2
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
tomlkit==0.13.3
typer==0.16.0
typing_extensions==4.14.1
wheel==0.45.1
